refactored some classes out of index.js
