+++ lib/index.js
@@ -1,27 +1,23 @@
 var canvas = document.getElementById('game');
 var context = canvas.getContext('2d');
+
+var lifeCounter = 0;
+const Cat = require('./cat')
+const Sushi = require('./sushi')
+const Trash = require('./trash')
+const Heart = require('./heart')
+
+var nyanCat = new Cat({context: context});
+
 var catDrawMinY = 50;
 var catDrawMaxY = 350;
-var nyanCat = new Cat();
-var heart1 = new Heart(500);
-var heart2 = new Heart(550);
-var heart3 = new Heart(600);
+var sushis = [];
+var trashes = [];
+var heart1 = new Heart(500, {context: context});
+var heart2 = new Heart(550, {context: context});
+var heart3 = new Heart(600, {context: context});
 var lastGenTime = 0;
 var points = 0;
-
-function Cat() {
-  this.image = document.getElementById("nyan-cat-image");
-  this.width = 100;
-  this.height = 100;
-  this.x = 10;
-  this.y = catDrawMinY;
-  this.context = context;
-}
-
-Cat.prototype.draw = function () {
-  this.context.drawImage(this.image, this.x, this.y);
-  return this;
-}
 
 $(document).on('keydown', function(event){
   if (event.keyCode === 38) {
@@ -31,72 +27,6 @@ $(document).on('keydown', function(event){
     nyanCat.y = Math.min(catDrawMaxY, nyanCat.y + nyanCat.height);
   }
 });
-
-function Sushi() {
-  var rowsForSprites = [70, 170, 270, 370];
-  var sushiImages = ['egg-roll', 'roe-roll', 'fish-roll'];
-  this.image = document.getElementById(sushiImages[Math.floor(Math.random() * sushiImages.length)]);
-  this.width = 70;
-  this.height = 58;
-  this.x = 600;
-  this.y = rowsForSprites[Math.floor(Math.random() * rowsForSprites.length)];
-  this.context = context;
-}
-
-function Trash() {
-  var rowsForTrash = [70, 170, 270, 370];
-  var trashImages = ['kawaii-poop', 'kawaii-toaster'];
-  this.image = document.getElementById(trashImages[Math.floor(Math.random() * trashImages.length)]);
-  this.width = 70;
-  this.height = 58;
-  this.x = 600;
-  this.y = rowsForTrash[Math.floor(Math.random() * rowsForTrash.length)];
-  this.context = context;
-}
-
-function Heart(x) {
-  this.image = document.getElementById("full-heart");
-  this.width = 60;
-  this.height = 60;
-  this.x = x;
-  this.y = 10;
-  this.context = context;
-}
-
-Sushi.prototype.draw = function() {
-  this.context.drawImage(this.image, this.x, this.y);
-  return this;
-}
-
-Sushi.prototype.move = function(sushis, index) {
-  this.x -= 3;
-  if (this.x < - 70){
-    clearObject(sushis, index);
-  }
-  return this;
-}
-
-var sushis = [];
-
-Trash.prototype.draw = function() {
-  this.context.drawImage(this.image, this.x, this.y);
-  return this;
-}
-
-Trash.prototype.move = function(trashes, index) {
-  this.x -= 3;
-  if (this.x < - 70){
-    clearObject(trashes, index);
-  }
-  return this;
-}
-
-var trashes = [];
-
-Heart.prototype.draw = function() {
-  this.context.drawImage(this.image, this.x, this.y);
-  return this;
-}
 
 requestAnimationFrame(function gameLoop() {
   context.clearRect(0, 0, canvas.width, canvas.height);
@@ -167,15 +97,13 @@ function makeObject() {
   var number = Math.random();
 
   if (number > 0.5) {
-    var sushi = new Sushi();
+    var sushi = new Sushi({context: context});
     sushis.push(sushi);
   } else {
-    var trash = new Trash();
+    var trash = new Trash({context: context});
     trashes.push(trash);
   }
 }
-
-var lifeCounter = 0;
 
 function loseHeart() {
   var hearts = [heart1, heart2, heart3];
